// src/api/types/course.type.ts
export enum CourseLabel {
  NEW = "NEW",
  HOT = "HOT",
  BEST_SELLER = "BEST_SELLER",
  FEATURED = "FEATURED",
  NONE = "NONE"
}

export interface Course {
  id: string;
  title: string;
  slug: string;
  shortDescription: string;
  thumbnail: string;
  difficulty: string;
  label?: string;
  totalLesson: number;
  owner: {
    id: string;
    email: string;
    fullName: string;
  }
  pricing: {
    regular: number;
    discounted?: number;
  };
  rating: {
    avg: number;
    count: number;
  };
  enrollmentCnt: number;
  status: EStatusCourse;
  discountedPrice: number,
  regularPrice: number,
  isFree: boolean;
  category: {
    id: string;
    title: string;
    slug: string;
  }
}